xfree86: ddc: move public SDK headers to include/
