+++ hw/xfree86/common/xf86Helper.c
@@ -41,6 +41,7 @@
 #include "dix/dix_priv.h"
 #include "dix/input_priv.h"
 #include "include/extinit.h"
+#include "include/xf86DDC.h"
 #include "mi/mi_priv.h"
 #include "os/log_priv.h"
 #include "os/osdep.h"
@@ -57,7 +58,6 @@
 #include "xf86_OSlib.h"
 #include "micmap.h"
 #include "xf86Bus.h"
-#include "xf86DDC.h"
 #include "xf86Xinput_priv.h"
 #include "xf86InPriv.h"
 #include "xf86Config.h"
