+++ hw/xfree86/common/xf86Init.c
@@ -52,6 +52,7 @@
 #include "config/hotplug_priv.h"
 #include "dix/input_priv.h"
 #include "dix/screenint_priv.h"
+#include "include/xf86DDC.h"
 #include "include/xorgVersion.h"
 #include "mi/mi_priv.h"
 #include "os/cmdline.h"
@@ -78,7 +79,6 @@
 #include "xf86cmap.h"
 #include "mipointer.h"
 #include "xf86Extensions.h"
-#include "xf86DDC.h"
 #include "xf86Xinput.h"
 #include "xf86InPriv.h"
 #include "xf86Crtc.h"
