+++ hw/xfree86/ddc/ddc.c
@@ -11,12 +11,12 @@
  */
 #include <xorg-config.h>
 
+#include "include/xf86DDC.h"
 #include "os/osdep.h"
 
 #include "misc.h"
 #include "xf86.h"
 #include "xf86_OSproc.h"
-#include "xf86DDC.h"
 #include <string.h>
 #include "edid_priv.h"
