+++ hw/xfree86/ddc/meson.build
@@ -11,5 +11,3 @@ xorg_ddc = static_library('xorg_ddc',
     dependencies: common_dep,
     c_args: xorg_c_args,
 )
-
-install_data(['edid.h', 'xf86DDC.h'], install_dir: xorgsdkdir)
