+++ hw/xfree86/ddc/xf86DDC_priv.h
@@ -6,7 +6,7 @@
 #ifndef _XSERVER_XF86_DDC_PRIV_H
 #define _XSERVER_XF86_DDC_PRIV_H
 
-#include "xf86DDC.h"
+#include "include/xf86DDC.h"
 #include "edid_priv.h"
 
 /*
