+++ hw/xfree86/modes/xf86DiDGA.c
@@ -21,8 +21,9 @@
  */
 #include <xorg-config.h>
 
+#include "include/xf86DDC.h"
+
 #include "xf86.h"
-#include "xf86DDC.h"
 #include "xf86_OSproc.h"
 #include "dgaproc.h"
 #include "dgaproc_priv.h"
