+++ hw/xfree86/modes/xf86RandR12.c
@@ -27,13 +27,13 @@
 #include "dix/input_priv.h"
 #include "dix/screenint_priv.h"
 #include "include/extinit.h"
+#include "include/xf86DDC.h"
 
 #include "xf86.h"
 #include "os.h"
 #include "globals.h"
 #include "xf86Modes.h"
 #include "xf86Priv.h"
-#include "xf86DDC.h"
 #include "mipointer.h"
 #include "windowstr.h"
 #include "inputstr.h"
