+++ include/meson.build
@@ -433,6 +433,7 @@ if build_xorg
             'dri2.h',
             'dri3.h',
             'dristruct.h',
+            'edid.h',
             'events.h',
             'exa.h',
             'exevents.h',
@@ -504,6 +505,7 @@ if build_xorg
             'xaarop.h',
             'xf86.h',
             'xf86cmap.h',
+            'xf86DDC.h',
             'xf86fbman.h',
             'xf86Opt.h',
             'xf86Optionstr.h',
